updated git ignore. Not sure how but in one of the merges it was modified and was missing node modules and the expo exec. Removed the 'explore' tab since it was unecessary. Renamed the homescreen (original index) to home.tsx. Made the login screen new index, so users have to login first
