+++ .gitignore
@@ -1,9 +1,13 @@
+Learn more https://docs.github.com/en/get-started/getting-started-with-git/ignoring-files
 
-# @generated expo-cli sync-2b81b286409207a5da26e14c78851eb30d8ccbdb
-# The following patterns were generated by expo-cli
+dependencies
+node_modules/
 
+Expo
+.expo/
+dist/
+web-build/
 expo-env.d.ts
-<<<<<<< HEAD
 
 Native
 .kotlin/
@@ -31,8 +35,3 @@ local env files
 
 typescript
 *.tsbuildinfo
-
-=======
-# @end expo-cli
-node_modules
->>>>>>> parent of d0b406a9 (Merge branch 'main' into KNBranch)
